Bug fixes and implementation of opacity control
git-svn-id: https://ncwms.svn.sourceforge.net/svnroot/ncwms/trunk@82 f020452a-5e1f-0410-8fa4-ae2b6fb27d5e

+++ src/java/uk/ac/rdg/resc/ncwms/graphics/PicMaker.java
@@ -28,7 +28,7 @@ public abstract class PicMaker
     protected float scaleMin;
     protected float scaleMax;
     // The percentage opacity of the picture
-    protected int opacity;
+    protected float opacity;
     // The fill value of the data.  Set to GmapsDataReader.FILL_VALUE by default
     protected float fillValue;
